refactor: complete post-v0.20 quality cleanup
