+++ crates/nose-cli/src/surfaces.rs
@@ -1,11 +1,15 @@
-use rayon::prelude::*;
 use rustc_hash::FxHashSet;
 
-use crate::source_lines::FileLineCache;
-
+mod declarations;
 mod generated;
 mod generated_paths;
 
+#[cfg(test)]
+pub(crate) use declarations::span_is_declarations;
+use declarations::{
+    declaration_only_type_contract_families, declaration_run_families,
+    family_declaration_only_type_contract, family_declaration_run,
+};
 use generated::{generated_source_indexes, GeneratedSourceIndexes};
 #[cfg(test)]
 pub(crate) use generated::{
@@ -150,279 +154,6 @@ pub(crate) fn family_actionability_reason(
     } else {
         family.actionability_reason()
     }
-}
-
-fn family_declaration_run(
-    family: &nose_detect::RefactorFamily,
-    overrides: &SurfaceOverrides,
-) -> bool {
-    overrides
-        .declaration_run_families
-        .contains(&family_handle(family))
-}
-
-fn family_declaration_only_type_contract(
-    family: &nose_detect::RefactorFamily,
-    overrides: &SurfaceOverrides,
-) -> bool {
-    overrides
-        .declaration_only_type_contract_families
-        .contains(&family_handle(family))
-}
-
-/// Apply the typed product form of `declaration-only-type.v1` frozen by #841.
-///
-/// Every member needs an un-sliced type-unit location plus positive whole-unit,
-/// type-only, declaration-only proof. The location-kind check prevents a
-/// pair-local connected witness from reusing its enclosing unit's origin after
-/// the actionable span has been narrowed to a block.
-/// The taxonomy's abstract `runtime` domain maps to the IL's `Imperative`
-/// domain and explicit runtime-value/validation flags. Data/implementation
-/// domains and every reusable/default/extension body signal are also defensive
-/// disqualifiers. Because every condition is positive and all-member, unknown,
-/// partial, mixed, default-body, extension, enum, and schema origins remain on
-/// their ranked surface.
-fn declaration_only_type_contract_families(
-    families: &[nose_detect::RefactorFamily],
-) -> FxHashSet<FamilyHandle> {
-    families
-        .iter()
-        .filter(|family| is_declaration_only_type_contract(family))
-        .map(family_handle)
-        .collect()
-}
-
-fn is_declaration_only_type_contract(family: &nose_detect::RefactorFamily) -> bool {
-    use nose_il::{
-        SourceGranularity, UnitBodyKind, UnitDomain, UnitEvidenceFlag, UnitKind, UnitSubkind,
-    };
-
-    !family.locations.is_empty()
-        && family.locations.iter().all(|location| {
-            let origin = location.origin;
-            location.kind == UnitKind::Class
-                && !location.is_fragment
-                && origin.has_domain(UnitDomain::TypeContract)
-                && origin
-                    .domains
-                    .iter()
-                    .all(|domain| domain == UnitDomain::TypeContract)
-                && matches!(
-                    origin.subkind,
-                    UnitSubkind::InterfaceTraitProtocol
-                        | UnitSubkind::TypeAlias
-                        | UnitSubkind::DefinedType
-                )
-                && origin.body_kind == UnitBodyKind::DeclarationOnly
-                && origin.source_granularity == SourceGranularity::WholeUnit
-                && origin.has_evidence(UnitEvidenceFlag::DeclarationOnly)
-                && origin.has_evidence(UnitEvidenceFlag::TypeOnly)
-                && !origin.evidence_flags.iter().any(|flag| {
-                    matches!(
-                        flag,
-                        UnitEvidenceFlag::HasRuntimeBody
-                            | UnitEvidenceFlag::HasReusableBody
-                            | UnitEvidenceFlag::RuntimeValue
-                            | UnitEvidenceFlag::RuntimeValidation
-                            | UnitEvidenceFlag::HasDefaultBody
-                            | UnitEvidenceFlag::ProtocolExtension
-                            | UnitEvidenceFlag::ConcreteTypeExtension
-                            | UnitEvidenceFlag::ConstrainedExtension
-                            | UnitEvidenceFlag::InterfaceDefaultMethod
-                            | UnitEvidenceFlag::InterfaceStaticMethod
-                            | UnitEvidenceFlag::InterfacePrivateMethod
-                    )
-                })
-        })
-}
-
-/// Classify the mechanically-decidable declaration runs in `families`.
-///
-/// A *declaration run* is a family whose every member span consists solely of
-/// import/include/use/re-export declarations (plus blank lines and full-line
-/// comments). The duplication is real — the syntax channel is right that the
-/// lines match — but the language mandates these declarations per file, so no
-/// extraction exists and no judgment is owed (design.md: provable
-/// non-actionability is the detector's job, not the consumer's).
-///
-/// Fail-open by construction: any line not provably part of a declaration, an
-/// unsupported extension, an unreadable span, or an unclosed multi-line
-/// statement keeps the family on its ranked surface. Misclassifying a real
-/// finding is the error class this guards against; missing an import run is
-/// only a ranking nuisance.
-fn declaration_run_families(families: &[nose_detect::RefactorFamily]) -> FxHashSet<FamilyHandle> {
-    // Three passes (coevo s4 perf packet): a cheap serial prescreen picks the
-    // candidate families, the unique candidate files parse in PARALLEL (the
-    // serial per-file AST parse cost +29% wall on sympy), and the final pass
-    // classifies against the shared facts.
-    let mut lines = FileLineCache::default();
-    let mut candidates: Vec<&nose_detect::RefactorFamily> = Vec::new();
-    let mut wanted: std::collections::BTreeSet<String> = std::collections::BTreeSet::new();
-    for f in families {
-        if !declaration_run_candidate(f) {
-            continue;
-        }
-        let pass = f.locations.iter().all(|l| {
-            lines
-                .whole(&l.file)
-                .is_some_and(|all| declaration_prescreen(all, l.start_line, l.end_line))
-        });
-        if pass {
-            candidates.push(f);
-            wanted.extend(f.locations.iter().map(|l| l.file.clone()));
-        }
-    }
-    let facts: std::collections::HashMap<String, Option<nose_frontend::DeclarationFacts>> = wanted
-        .into_iter()
-        .collect::<Vec<_>>()
-        .into_par_iter()
-        .map(|file| {
-            let parsed = std::path::Path::new(&file)
-                .extension()
-                .and_then(|e| e.to_str())
-                .and_then(|ext| {
-                    let src = std::fs::read_to_string(&file).ok()?;
-                    nose_frontend::declaration_facts(ext, &src)
-                });
-            (file, parsed)
-        })
-        .collect();
-    candidates
-        .iter()
-        .filter(|f| {
-            f.locations
-                .iter()
-                .all(|l| declaration_run_span(l, &mut lines, &facts))
-        })
-        .map(|f| family_handle(f))
-        .collect()
-}
-
-fn declaration_run_candidate(family: &nose_detect::RefactorFamily) -> bool {
-    !family.locations.is_empty()
-        && family
-            .witness
-            .as_ref()
-            .is_some_and(|w| w.kind == "copy-paste-run")
-        && family.locations.iter().all(|l| {
-            declaration_candidate_lang(&l.lang)
-                && l.kind == nose_il::UnitKind::Block
-                && l.name.is_none()
-                && l.end_line.saturating_sub(l.start_line) <= DECLARATION_SPAN_CAP
-        })
-}
-
-fn declaration_candidate_lang(lang: &str) -> bool {
-    matches!(lang, "javascript" | "typescript")
-}
-
-/// An import run longer than this is implausible; skip the read and fail open.
-const DECLARATION_SPAN_CAP: u32 = 80;
-
-fn declaration_run_span(
-    loc: &nose_detect::Loc,
-    lines: &mut FileLineCache,
-    facts: &std::collections::HashMap<String, Option<nose_frontend::DeclarationFacts>>,
-) -> bool {
-    if loc.end_line.saturating_sub(loc.start_line) > DECLARATION_SPAN_CAP {
-        return false;
-    }
-    let Some(Some(facts)) = facts.get(&loc.file) else {
-        return false;
-    };
-    let Some(all) = lines.whole(&loc.file) else {
-        return false;
-    };
-    span_is_declarations(facts, all, loc.start_line, loc.end_line)
-}
-
-/// Cheap starter check before the AST parse. Comment lines are transparent;
-/// the first content line must begin like wiring. False negatives only fail
-/// open (the family keeps its ranked surface), so this can never misclassify.
-fn declaration_prescreen(all: &[String], start: u32, end: u32) -> bool {
-    const STARTERS: &[&str] = &[
-        "import",
-        "from ",
-        "use ",
-        "pub use ",
-        "pub mod ",
-        "pub extern ",
-        "pub(",
-        "#include",
-        "#pragma",
-        "package ",
-        "require",
-        "export ",
-        "extern ",
-        "mod ",
-    ];
-    let end = (end as usize).min(all.len());
-    if start == 0 || start as usize > end {
-        return false;
-    }
-    for line in &all[start as usize - 1..end] {
-        // A leading UTF-8 BOM is invisible to the AST classifier (it strips
-        // one) — the prescreen must too, or a BOM'd first import never reaches
-        // the parse (coevo S4-C3).
-        let t = line.trim_start_matches('\u{feff}').trim_start();
-        if t.is_empty() || t.starts_with("//") || t.starts_with("/*") {
-            continue;
-        }
-        if t.starts_with('#') && !t.starts_with("#include") && !t.starts_with("#pragma") {
-            continue;
-        }
-        // A span may begin INSIDE a multi-line import (specifier list or its
-        // closer) — the AST node covers those lines, so let the parse decide.
-        if t.starts_with('}') || t.starts_with(')') {
-            return true;
-        }
-        if t.chars()
-            .all(|c| c.is_alphanumeric() || matches!(c, '_' | '$' | ',' | ' ' | '.'))
-        {
-            return true;
-        }
-        // CommonJS wiring needs the call, not just the keyword.
-        for head in ["const ", "let ", "var "] {
-            if t.starts_with(head) {
-                return t.contains("= require(");
-            }
-        }
-        return STARTERS.iter().any(|s| t.starts_with(s));
-    }
-    false
-}
-
-/// The line rule over AST facts: every line in the span must be blank, a
-/// comment, or part of a declaration statement; a single code-poisoned line
-/// (any named leaf outside declarations/comments — `import os; evil()` puts
-/// `evil()`'s leaves on the import's line) disqualifies the span; and at
-/// least one declaration line must be present.
-pub(crate) fn span_is_declarations(
-    facts: &nose_frontend::DeclarationFacts,
-    all: &[String],
-    start: u32,
-    end: u32,
-) -> bool {
-    let end = (end as usize).min(all.len()) as u32;
-    if start == 0 || start > end {
-        return false;
-    }
-    let mut any = false;
-    for line_no in start..=end {
-        if facts.is_code_line(line_no) {
-            return false;
-        }
-        if facts.is_declaration_line(line_no) {
-            any = true;
-            continue;
-        }
-        if facts.is_comment_line(line_no) || all[line_no as usize - 1].trim().is_empty() {
-            continue;
-        }
-        // Uncovered non-blank content (stray tokens, mid-statement cuts).
-        return false;
-    }
-    any
 }
 
 fn family_all_generated_source(
