+++ crates/nose-frontend/Cargo.toml
@@ -25,7 +25,7 @@ rmp-serde = { workspace = true }
 # lowering is byte-identical to go 0.23 — required, else the go Raw ratio explodes.
 # See #403. VALIDATE any bump on a `cargo clean` release build — incremental builds
 # reuse stale cc-compiled grammar objects and pass falsely (#399/#400/#401 → #402).
-tree-sitter = "0.25"
+tree-sitter = "0.26"
 tree-sitter-python = "0.25"
 tree-sitter-javascript = "0.25"
 tree-sitter-typescript = "0.23"
